feat: startup configuration (#18)
* Allow coordinator to be set manually

* Add manual coordinator test

* Implement across message handler

* Send signature to all relayers

* Implement signature cache to store generated signatures

* Speed up process coordination

* Implement signature cache to store  generated signatures

* Add signature cache tests

* Allow for coordinator to notify other relayers of an across message

* Reduce tss timeout

* Implement calculation of the unlock hash that matches on-chain implementation

* Add across message handler tests

* Update mocks to maintained version

* Fix test race condition

* Convert abi to constant

* Use source chain id property from across data

* Use message destination as across deposit source

* Use coordinator from the host in message handler

* Send across message result over the err chanel

* Return api errors as json

* Implement api serve function

* Add signing handler tests

* Ignore valid json encode errors

* Enforce protocol type on the signing API

* Update signing url

* Implement signature cache subscribe

* Implement status sse endpoint

* Bump action cache action

* Set chain id as uint64

* Deprecate propstore

* Add across pool address to config

* Make signature cache watch a separate function

* Add api addr to config

* Implement startup configuration

* Remove uploader config

* Bump golang version

* Add example config

* Fix app startup

* Fix across message infinite loop

* Fix across tests

* Make test less flaky

* Remove print

* Remove unnecessary waits

* Lint

* Ignore irrelevant coordinator errors

* Try to execute tss process in spite of not notifying everyone

* Fix coordinator error assignemnt

* Add manual timeout to keygen

* Increase coordintor timeout in tests

* Increase tss process timeout

+++ config/chain/config_test.go
@@ -8,7 +8,7 @@ import (
 )
 
 func TestValidateConfig(t *testing.T) {
-	var id uint8 = 1
+	var id uint64 = 1
 	valid := GeneralChainConfig{
 		Name:     "chain",
 		Id:       &id,
